requirements/base.txt
Django==1.8.1
coverage==3.7.1
Django==1.8.1
django-appconf==1.0.1
django-ckeditor==4.4.7
django-image-cropping==1.0.1
easy-thumbnails==2.2
Pillow==2.8.2
psycopg2==2.6
selenium==2.45.0
six==1.9.0
Unidecode==0.4.18
